refactor: decouple project index dependencies
